Add color change functionality to user profile
This commit adds the ability for users to change their favorite color on their profile. The `validateColor` middleware was added to ensure that the provided color is one of the allowed colors. The `changeColor` controller function was also added to handle the color change request. The `User` model was updated to use the `colors` helper module for the `favColor` enum. The `profile.js` route file was updated to include the new color change route.